backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.28
alembic>=1.13.1
asyncpg>=0.29.0
httpx>=0.27.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
apscheduler>=3.10.0
openpyxl>=3.1.0
